refactor: PaymentClient의 승인 메소드 반환값 도메인 객체로 변경
