+++ src/main/java/roomescape/payment/domain/PaymentClient.java
@@ -1,8 +1,7 @@
 package roomescape.payment.domain;
 
 import roomescape.payment.dto.request.PaymentConfirmRequest;
-import roomescape.payment.dto.response.PaymentConfirmResponse;
 
 public interface PaymentClient {
-    PaymentConfirmResponse confirm(PaymentConfirmRequest request);
+    ConfirmedPayment confirm(PaymentConfirmRequest request);
 }
