+++ src/main/java/roomescape/reservation/event/ReservationSavedEvent.java
@@ -1,6 +1,7 @@
 package roomescape.reservation.event;
 
-import roomescape.payment.domain.Payment;
+import roomescape.payment.domain.ConfirmedPayment;
+import roomescape.reservation.domain.Reservation;
 
-public record ReservationSavedEvent(Payment payment) {
+public record ReservationSavedEvent(Reservation reservation, ConfirmedPayment confirmedPayment) {
 }
