+++ src/test/java/roomescape/common/StubPaymentClient.java
@@ -1,13 +1,13 @@
 package roomescape.common;
 
+import roomescape.payment.domain.ConfirmedPayment;
 import roomescape.payment.domain.PaymentClient;
 import roomescape.payment.dto.request.PaymentConfirmRequest;
-import roomescape.payment.dto.response.PaymentConfirmResponse;
 
 public class StubPaymentClient implements PaymentClient {
 
     @Override
-    public PaymentConfirmResponse confirm(PaymentConfirmRequest request) {
-        return new PaymentConfirmResponse("paymentKey", "orderId", 10);
+    public ConfirmedPayment confirm(PaymentConfirmRequest request) {
+        return new ConfirmedPayment("paymentKey", "orderId", 10);
     }
 }
